fix(api): hash password on password update

+++ services/api-server/src/users/users.controller.ts
@@ -19,6 +19,7 @@ import { GetUserByInviteCodeDTO } from './dto/getUserByInviteCode.dto';
 import { Request } from 'express';
 import { SetupProfileDTO } from './dto/setupProfile.dto';
 import { StoreService } from 'src/store/store.service';
+import { genSaltSync, hashSync } from 'bcryptjs';
 
 @Controller('users')
 export class UsersController {
@@ -57,9 +58,10 @@ export class UsersController {
   async getMe(@Req() req: Request) {
     try {
       const userId = req.user?.['sub'];
-      const user = await this.usersService.findById(userId);
+      const user = await this.usersService.findById(userId!);
       return { user };
     } catch (error) {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
       return { message: error.message };
     }
   }
@@ -73,11 +75,13 @@ export class UsersController {
     try {
       const userId = req.user?.['sub'];
 
+      const hash = hashSync(body.password, genSaltSync(10));
+
       // 1. Update user profile
-      await this.usersService.update(userId, {
+      await this.usersService.update(userId!, {
         firstName: body.firstName,
         lastName: body.lastName,
-        password: body.password,
+        password: hash,
         image: body.profileImage,
         onboardingCompleted: true,
         inviteCode: null,
@@ -85,7 +89,7 @@ export class UsersController {
         tosAcceptedAt: new Date(),
       });
 
-      const user = await this.usersService.findById(userId)
+      const user = await this.usersService.findById(userId!);
 
       const store = await this.storeService.createStoreWithUserRole({
         name: body.storeName,
@@ -96,7 +100,7 @@ export class UsersController {
         locationLng: body.storeLocationLng,
         customerPaysFees: body.customerPaysFees,
         partnerId: user.partnerId,
-        userId,
+        userId: userId!,
       });
 
       return { user, store };
@@ -110,8 +114,10 @@ export class UsersController {
   async deleteUser(@Req() req: Request) {
     try {
       const userId = req.user?.['sub'];
-      await this.usersService.delete(userId);
-      return { message: 'User deleted successfully' };
+      if (userId) {
+        await this.usersService.delete(userId);
+        return { message: 'User deleted successfully' };
+      } else throw new Error('');
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
